[Refactoring] Remove big-integer library (#6294)

+++ src/api/gramjs/updates/UpdatePts.ts
@@ -1,15 +1,13 @@
-import type { BigInteger } from 'big-integer';
-
 export class LocalUpdatePts {
   constructor(public pts: number, public ptsCount: number) {}
 }
 
 export class LocalUpdateChannelPts {
-  constructor(public channelId: BigInteger, public pts: number, public ptsCount: number) {}
+  constructor(public channelId: bigint, public pts: number, public ptsCount: number) {}
 }
 
 export type UpdatePts = LocalUpdatePts | LocalUpdateChannelPts;
 
-export function buildLocalUpdatePts(pts: number, ptsCount: number, channelId?: BigInteger) {
+export function buildLocalUpdatePts(pts: number, ptsCount: number, channelId?: bigint) {
   return channelId ? new LocalUpdateChannelPts(channelId, pts, ptsCount) : new LocalUpdatePts(pts, ptsCount);
 }
